// src/dto/accountbook-dto.ts
export class AccountbookDto {
    serialNo: number;
    userId: string;
    dateYmd: string;
    kanjoNo: number;
    attribute: string;
    amount: number;
    content: string;
    detail: string;
    withwho: string;
    furikaeKbn: string;
    sourceCd: string;
    sourceSerial: number;
    kyouseiKbn: string;
    torikeshiKbn: string;
    deleteKbn: string;
    insertUserId: string;
    updateUserId: string;
    insertTs: Date;
    updateTs: Date;
}